comments added to the code

+++ Develop/Assets/js/index.js
@@ -1,24 +1,24 @@
-let APIkey = '6a603a1b7695d9e567a903c3edc9723d';
-let otherAPI = '408195a1df5c96395e90601ba41a5167';
+let APIkey = '6a603a1b7695d9e567a903c3edc9723d'; //apikey for current weather
+let otherAPI = '408195a1df5c96395e90601ba41a5167';   //apikey for forecasted weather
 let city;
 let city2;
-let url = 'https://api.openweathermap.org/data/2.5/weather?q='+city+'&appid='+APIkey;
-let urlForecast = 'https://api.openweathermap.org/data/2.5/forecast?q='+city+'&appid='+otherAPI;
+let url = 'https://api.openweathermap.org/data/2.5/weather?q='+city+'&appid='+APIkey; //url to call the current weather api
+let urlForecast = 'https://api.openweathermap.org/data/2.5/forecast?q='+city+'&appid='+otherAPI; //url to call the forecasted weather api
 let btnEl = document.getElementById('btn')
 let inputEl = document.querySelector('input')
 let cities = document.getElementById('cities')
 let storedCity = JSON.parse(localStorage.getItem("cities"))
 let array=[];
 
-if(storedCity!== null){
+if(storedCity!== null){  //if stored values not null, display the cities stored 
     array=storedCity;
-    for(let i=0; i<storedCity.length; i++){
-        cities.innerHTML+=`<button id="${storedCity[i]}">${storedCity[i]}</button>`
+    for(let i=0; i<storedCity.length; i++){  //create a button for each of the cities
+        cities.innerHTML+=`<button id="${storedCity[i]}">${storedCity[i]}</button>` 
     }  
 }
     
-cities.addEventListener("click", function(event){
-    if(event.target !== event.currentTarget){
+cities.addEventListener("click", function(event){ //if one button created for the cities is clicked will make a call for the 2 api
+    if(event.target !== event.currentTarget){ //only if one of the buttons is clicked
         city = event.target.id; 
         url = 'https://api.openweathermap.org/data/2.5/weather?q='+city+'&appid='+APIkey+'&units=metric';
         urlForecast = 'https://api.openweathermap.org/data/2.5/forecast?q='+city+'&appid='+otherAPI+'&units=metric';
@@ -27,7 +27,7 @@ cities.addEventListener("click", function(event){
     }
 })
 
-$('#t').append(dayjs().format(' M/D/YYYY'))
+$('#t').append(dayjs().format(' M/D/YYYY'))  //displaying current date and the date for the next five days
 $('#t1').text(dayjs().add(1, 'day').format('M/D/YYYY'))
 $('#t2').text(dayjs().add(2, 'day').format('M/D/YYYY'))
 $('#t3').text(dayjs().add(3, 'day').format('M/D/YYYY'))
@@ -39,18 +39,18 @@ var responseText = document.getElementById('response-text');
 function today(url) {
   fetch(url)
     .then(function (response) {
-        if (response.ok){
+        if (response.ok){  //if the call for the api is okay 
             response.json().then(function (data) {
-                city2= city.charAt(0).toUpperCase() + city.slice(1);
+                city2= city.charAt(0).toUpperCase() + city.slice(1);  //this is made to capitalize the first letter in the string
                 $('#t').text(city2+dayjs().format(' M/D/YYYY'))
                 let iconCode = data.weather[0].icon;
                 let iconUrl="http://openweathermap.org/img/w/" + iconCode + ".png";        
-                $('#icon').attr('src', iconUrl)
+                $('#icon').attr('src', iconUrl)   //takes the icon given by the api
                 $('#icon').attr('alt', "weather icon forecast")
                 $('#temp').text('Temp: '+data.main.temp+' °C')
                 $('#wind').text('Wind: '+data.wind.speed+' m/s')
                 $('#hum').text('Humidity: '+data.main.humidity+' %')
-                if(!array.includes(city)){
+                if(!array.includes(city)){ //if the city entered is not already in the array will push it in and then store it and display it with a button
                     array.push(city)
                     localStorage.setItem("cities", JSON.stringify(array))
                     storedCity = JSON.parse(localStorage.getItem("cities"))
@@ -67,24 +67,24 @@ function today(url) {
 function forecast(urlForecast){
     fetch(urlForecast)
     .then(function (response) {
-        if (response.ok){
+        if (response.ok){  //if the respone is ok
             response.json().then(function (data) {
 
                 let prevHour;
                 let nextHour;
                 let day;
                 
-                let currentHour = dayjs().add(1, 'day').format('HH');
-                
-                for(let i=0; i<data.list.length-1;i++){
-
+                let currentHour = dayjs().format('HH');
+                //from our data we receive an array with 40 forecasted values
+                for(let i=0; i<data.list.length-1;i++){  
+                    //data.list.dt_txt gives the day and the time back in a single string, so we slice it to have the 2 values separately
                     prevHour=data.list[i].dt_txt;
                     day = prevHour.slice(0,10);
-                    nextHour=data.list[i+1].dt_txt;
-                    let timeComp = (currentHour>=prevHour.slice(11,13) && currentHour<nextHour.slice(11,13))
+                    nextHour=data.list[i+1].dt_txt;  
+                    let timeComp = (currentHour>=prevHour.slice(11,13) && currentHour<nextHour.slice(11,13))  //compares is the current hour is between the last and the following forecasted value
 
-                    if(dayjs().add(1, 'day').format('YYYY-MM-D')===day  && timeComp){
-                        let forecast1 = i;
+                    if(dayjs().add(1, 'day').format('YYYY-MM-D')===day  && timeComp){ 
+                        let forecast1 = i; //we save the i value to take the elements for that iteration
                         let iconcode = data.list[forecast1].weather[0].icon;
                         let iconurl="http://openweathermap.org/img/w/" + iconcode + ".png";
                         $('#icon1').attr('src', iconurl)
@@ -141,7 +141,7 @@ function forecast(urlForecast){
     });
 }
 
-btnEl.addEventListener("click", function(){
+btnEl.addEventListener("click", function(){ //when button clicked we save the city value and call the apis
     if(inputEl.value){
         city = inputEl.value
         inputEl.value=''
